requirements.txt
streamlit>=1.30
numpy>=1.23
pandas>=1.5
plotly>=5.15
scikit-learn>=1.3
openpyxl>=3.1
pillow>=9.0
rdkit-pypi>=2022.9.5
pillow>=10.0